Refactor HLRIntake and SCIntake into a ClaimReviewIntake
