+++ app/models/concerns/asyncable.rb
@@ -1,0 +1,9 @@
+module Asyncable
+  extend ActiveSupport::Concern
+
+  private
+
+  def run_async?
+    !Rails.env.development? && !Rails.env.test?
+  end
+end
